Final Master Push! updated outdated readme file.
I reverted this branch to the latest push before the deadline so you can
see exactly see what we had accomplished on the "release date". We will
continue to update our game on the sprint 4 branch.
We had the game in a fully functional and shippable state before the
deadline, so we thought it would be important to show that. Although
ofcourse a game is never fully done and can always be tweaked, added
upon. which is exactly what we are going to do on the sprint 4 branch!